@@ -5,7 +5,7 @@
   "main": "build/index.js",
   "author": "Automattic, Inc.",
   "scripts": {
-    "prepublish": "make build"
+    "prepublish": "make standalone"
   },
   "contributors": [
     "Damian Suarez <rdsuarez@gmail.com>",
@@ -24,6 +24,14 @@
   ],
   "repository": "git@github.com:Automattic/wpcom.js.git",
   "license": "MIT",
+  "babel": {
+    "presets": [ "es2015", "stage-2" ]
+  },
+  "files": [
+    "build",
+    "History.md",
+    "README.md"
+  ],
   "dependencies": {
     "babel-runtime": "^6.9.2",
     "debug": "^2.2.0",
@@ -32,10 +40,13 @@
   },
   "devDependencies": {
     "babel-eslint": "^4.1.6",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-stage-2": "^6.11.0",
+    "babel-register": "^6.9.0",
     "eslint": "^1.10.1",
-    "mocha": "^2.3.4",
-    "n8-make": "^1.3.0",
-    "webpack": "^1.12.1",
+    "mocha": "^2.5.3",
+    "n8-make": "1.4.0",
+    "webpack": "^1.13.1",
     "wpcom-oauth-cors": "^1.0.0",
     "wpcom-proxy-request": "^2.0.0"
   }
