@@ -1,38 +1,19 @@
 module.exports = {
-	entry: __dirname + '/index.js',
+	entry: __dirname + '/build/index.js',
 
 	node: {
 		fs: 'empty'
 	},
 
 	output: {
-		path: __dirname + '/dist',
+		path: __dirname + '/build',
 		filename: 'wpcom.js',
 		libraryTarget: 'var',
 		library: 'WPCOM'
 	},
 
-	module: {
-		loaders: [
-			{
-				test: /\.js$/,
-				exclude: /node_modules/,
-				loader: 'babel-loader',
-				query: {
-					cacheDirectory: true,
-					optional: [ 'runtime' ]
-				}
-			},
-			{
-				test: /\.json$/,
-				exclude: /node_modules/,
-				loader: 'json-loader'
-			}
-		]
-	},
-
 	resolve: {
-		extensions: [ '', '.js', '.json' ]
+		extensions: [ '', '.js' ]
 	},
 
 	devtool: 'sourcemap'
